backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-ckeditor-5==0.2.17
django-contrib-comments==2.2.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-filter==25.1
django-fluent-comments==3.0
django-tag-parser==3.2
django-taggit==6.1.0
django-unfold==0.54.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
joblib==1.4.2
numpy==2.2.4
oauthlib==3.2.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-akismet==0.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
